Refactor API swagger documentaion
