+++ app/__init__.py
@@ -1,11 +1,9 @@
-from flasgger import Swagger
 from flask import Flask
 from flask_moment import Moment
 from flask_cors import CORS
 
-from app.models import User
 from instance.config import app_config
-from .template import template
+from swagger_ui.flask_swagger_ui import get_swaggerui_blueprint
 
 moment = Moment()
 
@@ -17,13 +15,19 @@ def create_app(config_name):
     app_config[config_name].init_app(app)
     CORS(app)
 
-    Swagger(app, template=template)
+    swagger_url = '/api/v1/docs'
+    api_url = 'swagger_doc.yml'
+
+    swaggerui_blueprint = get_swaggerui_blueprint(swagger_url, api_url)
+
     moment.init_app(app)
 
+    app.register_blueprint(swaggerui_blueprint, url_prefix=swagger_url)
+
     from app.api.v1.auth import auth as auth_blueprint
-    app.register_blueprint(auth_blueprint)
+    app.register_blueprint(auth_blueprint, url_prefix='/api/v1/')
 
     from app.api.v1.orders import orders as orders_blueprint
-    app.register_blueprint(orders_blueprint)
+    app.register_blueprint(orders_blueprint, url_prefix='/api/v1/')
 
     return app
