+++ app/api/v1/auth/__init__.py
@@ -1,5 +1,5 @@
 from flask import Blueprint
 
-auth = Blueprint('auth', __name__, url_prefix='/api/v1/')
+auth = Blueprint('auth', __name__)
 
 from . import views
